requirements.txt
nfcpy
gspread
pygame
requests
python-dotenv
google-auth